refactor: repo store, inline utility functions
