+++ ui/src/lib/auth.tsx
@@ -5,6 +5,7 @@ import {
   InMemoryCache,
   HttpLink,
   gql,
+  useQuery,
 } from "@apollo/client";
 import jwt_decode from "jwt-decode";
 
@@ -177,4 +178,26 @@ function useProvideAuth() {
     isSignedIn,
     hasToken,
   };
+}
+
+const PROFILE_QUERY = gql`
+  query Me {
+    me {
+      firstname
+      lastname
+      email
+      id
+      codeiumAPIKey
+    }
+  }
+`;
+
+//   avatar_url
+
+export default function useMe() {
+  /* eslint-disable no-unused-vars */
+  const { loading, data } = useQuery(PROFILE_QUERY, {
+    // fetchPolicy: "network-only",
+  });
+  return { loading, me: data?.me };
 }
