+++ ui/src/lib/utils/rich-schema.ts
