+++ ui/src/lib/utils/utils.ts
@@ -70,3 +70,44 @@ export const level2color = {
   // default: "rgba(255, 255, 255, 0.2)",
   default: "rgba(240,240,240,0.25)",
 };
+
+const yRemoteSelectionStyle = (clientID: string, color: string) => {
+  return `.yRemoteSelection-${clientID} 
+    { background-color: ${color}; opacity: 0.5;} `;
+};
+
+const yRemoteSelectionHeadStyle = (clientID: string, color: string) => {
+  return `.yRemoteSelectionHead-${clientID} {  
+        position: absolute;
+        border-left: ${color} solid 2px;
+        border-top: ${color} solid 2px;
+        border-bottom: ${color} solid 2px;
+        height: 100%;
+        box-sizing: border-box;}`;
+};
+
+const yRemoteSelectionHeadHoverStyle = (
+  clientID: string,
+  color: string,
+  name: string
+) => {
+  return `.yRemoteSelectionHead-${clientID}:hover::after { 
+        content: "${name}"; 
+        background-color: ${color}; 
+        box-shadow: 0 0 0 2px ${color};
+        border: 1px solid ${color};
+        color: white;
+        opacity: 1; }`;
+};
+
+export function addAwarenessStyle(
+  clientID: string,
+  color: string,
+  name: string
+) {
+  const styles = document.createElement("style");
+  styles.append(yRemoteSelectionStyle(clientID, color));
+  styles.append(yRemoteSelectionHeadStyle(clientID, color));
+  styles.append(yRemoteSelectionHeadHoverStyle(clientID, color, name));
+  document.head.append(styles);
+}
